requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
lxml
pydantic
pydantic-settings
pymongo
python-dotenv
scikit-learn
certifi